refactor: 검증 로직 추가

+++ src/main/java/com/example/meezy/bc/team/chat_room/application/service/DeleteChatRoomService.java
@@ -7,6 +7,7 @@ import com.example.meezy.bc.team.chat_room.domain.repository.ChatRoomRepository;
 import com.example.meezy.bc.team.team.application.service.exception.TeamNotFoundException;
 import com.example.meezy.bc.team.team.domain.Team;
 import com.example.meezy.bc.team.team.domain.repository.TeamRepository;
+import com.example.meezy.bc.team.team.domain.vo.TeamId;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,9 +25,9 @@ public class DeleteChatRoomService {
     @Transactional
     public void delete(UUID teamId, UUID chatRoomId){
         Team team = findTeamOrThrow(teamId);
-
         ChatRoom chatRoom = findChatRoomOrThrow(chatRoomId);
 
+        validateTeamOwnership(chatRoom, teamId);
         team.validateLeaderPermission(currentUserQuery.currentUser().userId());
 
         chatRoomRepository.delete(chatRoom);
@@ -40,5 +41,11 @@ public class DeleteChatRoomService {
     private ChatRoom findChatRoomOrThrow(UUID chatRoomId){
         return chatRoomRepository.findByChatRoomId_Value(chatRoomId)
                 .orElseThrow(ChatRoomNotFoundException::new);
+    }
+
+    private void validateTeamOwnership(ChatRoom chatRoom, UUID teamId) {
+        if (!chatRoom.getTeamId().equals(TeamId.of(teamId))) {
+            throw new ChatRoomNotFoundException();
+        }
     }
 }
